Block 1 — CI, lint & build hygiene foundation
- Add just check/test/lint recipes (CI parity, no hardware)
- Add GitHub Actions workflow (.github/workflows/ci.yml): Nix install,
  wasmi submodule over HTTPS, check/test/lint + full release-lto firmware
  link
- Fix all rustc warnings across app, desktop, firmware (lib+bin), and
  path-dependency crates (display_renderer, display_types, wasm_protocol,
  embedded_tools, esp32s3_embedded_tools, bq25895)
- Make clippy -D warnings clean on app (union feature set), desktop, and
  the firmware lib target
- Wire web import checks into the build (deno task build now runs
  check-imports; add deno task lint alias)
- DeviceConfig serde audit: named default functions for device_name /
  ap_password / known_wifi_networks (bare #[serde(default)] on String
  would deserialize to ""), with legacy-JSON deserialization tests
- wasmi submodule: drop stable vec_into_raw_parts from the feature gate
- Delete dead code: firmware/src/device.rs, firmware/src/native/ (never
  declared as a module — crate::native is the re-export of app::native),
  unused print_memory debug helper
- AGENTS.md: document the new recipes + CI, strike resolved items, add
  serde-default pitfall

+++ desktop/src/platform/display.rs
@@ -49,7 +49,7 @@ impl DesktopDisplayManager {
   /// `frame_buffer()` can expose it to the WebSocket remote-view handler.
   pub fn update_framebuffer(&self, pixels: &[u8]) {
     if let Ok(mut fb) = FRAMEBUFFER.lock() {
-      let buf = fb.get_or_insert_with(|| vec![0u8; (DISPLAY_WIDTH * DISPLAY_HEIGHT * 2) as usize]);
+      let buf = fb.get_or_insert_with(|| vec![0u8; DISPLAY_WIDTH * DISPLAY_HEIGHT * 2]);
       buf.copy_from_slice(pixels);
     }
   }
